+++ speeduino/corrections.h
@@ -44,7 +44,7 @@ extern byte activateMAPDOT; //The mapDOT value seen when the MAE was activated.
 extern byte activateTPSDOT; //The tpsDOT value seen when the MAE was activated.
 
 extern uint16_t AFRnextCycle;
-extern unsigned long knockStartTime;
+extern uint32_t knockStartTime;
 extern byte lastKnockCount;
 extern int16_t knockWindowMin; //The current minimum crank angle for a knock pulse to be valid
 extern int16_t knockWindowMax;//The current maximum crank angle for a knock pulse to be valid
