+++ speeduino/idle.h
@@ -26,7 +26,7 @@ struct StepperIdle
   int curIdleStep; //Tracks the current location of the stepper
   int targetIdleStep; //What the targeted step is
   volatile StepperStatus stepperStatus;
-  volatile unsigned long stepStartTime;
+  volatile uint32_t stepStartTime;
   byte lessAirDirection;
   byte moreAirDirection;
 };
