+++ speeduino/src/decoders/4g63.cpp
@@ -416,7 +416,7 @@ uint16_t getRPM_4G63(void)
     if (currentStatus.RPM < currentStatus.crankRPM)
     {
       int tempToothAngle;
-      unsigned long toothTime;
+      uint32_t toothTime;
 
       if (toothLastToothTime == 0 || toothLastMinusOneToothTime == 0)
       {
@@ -434,7 +434,7 @@ uint16_t getRPM_4G63(void)
         interrupts();
 
         toothTime *= 36;
-        tempRPM = ((unsigned long)tempToothAngle * (MICROS_PER_MIN / 10U)) / toothTime;
+        tempRPM = ((uint32_t)tempToothAngle * (MICROS_PER_MIN / 10U)) / toothTime;
         SetRevolutionTime((10UL * toothTime) / tempToothAngle);
         MAX_STALL_TIME = 366667UL; // 50RPM
       }
@@ -464,7 +464,7 @@ int getCrankAngle_4G63(void)
     //This is the current angle ATDC the engine is at. This is the last known
     //position based on what tooth was last 'seen'. It is only accurate to the
     //resolution of the trigger wheel (Eg 36-1 is 10 degrees)
-    unsigned long tempToothLastToothTime;
+    uint32_t tempToothLastToothTime;
     int tempToothCurrentCount;
     //Grab some variables that are used in the trigger code and assign them to temp variables.
 
@@ -472,7 +472,7 @@ int getCrankAngle_4G63(void)
 
     tempToothCurrentCount = toothCurrentCount;
     tempToothLastToothTime = toothLastToothTime;
-    unsigned long const lastCrankAngleCalc = micros();
+    uint32_t const lastCrankAngleCalc = micros();
 
     interrupts();
 
@@ -481,7 +481,7 @@ int getCrankAngle_4G63(void)
     crankAngle = toothAngles[tempToothCurrentCount - 1] + configPage4.triggerAngle;
 
     //Estimate the number of degrees travelled since the last tooth}
-    unsigned long const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
+    uint32_t const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
     crankAngle += timeToAngleIntervalTooth(elapsedTime);
 
     if (crankAngle >= 720)
